+++ motion-prime/config/base.rb
@@ -20,5 +20,7 @@ MotionPrime::Config.configure do |config|
     api.api_namespace = '/api'
     api.request_format = :form_data
     api.allow_redirect = false
+    api.allow_queue = false
   end
+  config.prime.cell_section.mixins = [Prime::CellSectionMixin]
 end